fix(network): точная (де)сериализация денег через BigDecimal
DTO хранили amount/balance как Double — BigDecimal.toDouble() на push и
Double при парсинге pull искажали десятичные суммы (домен, БД и бэкенд
уже точные: decimal). Введён BigDecimalTypeAdapter (читает BigDecimal из
числового токена, пишет toPlainString как JSON-число), общий Gson в
SerializationModule подключён к обоим Retrofit, DTO переведены на
BigDecimal, из мапперов убран toDouble(). Контракт не меняется — на
проводе остаётся JSON-число, бэкенд не трогаем. Добавлены тесты
адаптера и маппинга.

+++ core/data/src/main/java/soft/divan/financemanager/core/data/dto/AccountDto.kt
@@ -1,6 +1,7 @@
 package soft.divan.financemanager.core.data.dto
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class AccountDto(
     @SerializedName("id")
@@ -10,7 +11,7 @@ data class AccountDto(
     @SerializedName("name")
     val name: String,
     @SerializedName("balance")
-    val balance: Double,
+    val balance: BigDecimal,
     @SerializedName("currencyId")
     val currencyId: String,
     @SerializedName("createdAt")
